01-basic-agent/pyproject.toml
[project]
name = "01-basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.51.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
